move codeowner (#27)
# Description

Please include a summary of the change and which issue is fixed. Please
also
include relevant motivation and context. List any dependencies that are
required for this change.

# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Please
also note
any relevant details for your test configuration.

- [x] Current test suite

# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] Any dependent changes have been merged and published in downstream
modules